refactor(product_modify): product, productLike 구조 변경
1. productEntity -> product 로 엔티티 명칭 수정

2. product -> productLike  일대일
단방향 구조로 변경

3. 상품의 카테고리별로 맵핑되는 INT값 부여
