+++ src/main/java/com/zerobase/wishmarket/domain/product/model/type/ProductCategory.java
@@ -1,7 +1,14 @@
 package com.zerobase.wishmarket.domain.product.model.type;
 
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
 public enum ProductCategory {
-    CLOTHES, HOME_ELECTRONICS, TOY, IT_DEVICE, JEWELRY, FURNITURE, ETC
+    CLOTHES(0), HOME_ELECTRONICS(1), TOY(2),
+    IT_DEVICE(3), JEWELRY(4), FURNITURE(5),
+    ETC(6)
 
     ;
+
+    private final int categoryCode;
 }
