+++ src/main/java/com/zerobase/wishmarket/domain/product/repository/ProductRepository.java
@@ -1,10 +1,10 @@
 package com.zerobase.wishmarket.domain.product.repository;
 
-import com.zerobase.wishmarket.domain.product.model.entity.ProductEntity;
+import com.zerobase.wishmarket.domain.product.model.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
 }
