+++ src/test/java/com/zerobase/wishmarket/domain/product/exception/ProductErrorCodeTest.java
@@ -2,7 +2,7 @@ package com.zerobase.wishmarket.domain.product.exception;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import com.zerobase.wishmarket.domain.product.model.Product;
+import com.zerobase.wishmarket.domain.product.model.entity.Product;
 import com.zerobase.wishmarket.domain.product.repository.ProductRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
@@ -17,7 +17,7 @@ class ProductErrorCodeTest {
 
     //예외처리 테스트
     @Test
-    public void productFindTest() {
+    public void productExceptionTest() {
 
         assertThrows(ProductNotFoundException.class, () -> {
             Optional<Product> product = productRepository.findById(1L);
@@ -26,4 +26,5 @@ class ProductErrorCodeTest {
             }
         });
     }
+
 }
